migrate another pathological case
